requirements.txt
websockets
aiohttp
streamlit
pandas
plotly
requests
rich
python-dotenv
